feat: add insight2boardCfg to sb
