+++ packages/auto-chart/package.json
@@ -83,7 +83,7 @@
   "limit-size": [
     {
       "path": "dist/index.min.js",
-      "limit": "756 Kb",
+      "limit": "800 Kb",
       "gzip": true
     },
     {
