+++ packages/smart-board/package.json
@@ -87,7 +87,7 @@
     },
     {
       "path": "dist/index.min.js",
-      "limit": "512 Kb"
+      "limit": "384 Kb"
     }
   ]
 }
